Couple of things...
Forgot to

npm install aws-amplify when starting this project. I had thought that was the amplify cli which I already had. It wasn't it was the library to use here...

Trying to show Protected component fails because no 'history' property exists on props... Actually makes sense as we haven't added routing yet ! ! !

Maybe show these sorts of errors as WE ALL build it, slowly, but surely ! ! !